requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
python-dotenv==1.0.1
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1
channels==4.1.0
daphne==4.2.1
google-auth==2.34.0
requests==2.32.3
stripe==10.4.0
